+++ src/components/Filter/index.js
@@ -1,0 +1,6 @@
+import { FilterButton } from "./FilterCategory";
+import FilterProvider, { useFilter } from "./FilterProvider";
+import { FilterContainer } from "./FilterContainer";
+
+// Slug for the components folder src/components/Filter
+export { FilterButton, FilterProvider, useFilter, FilterContainer };
